Canged microcontroller centric unsigned 8-bit integer to a standard int for ease of understanding

+++ Arduino/djo_RP2040CAN_quickJoystick/djo_RP2040CAN_quickJoystick.ino
@@ -23,7 +23,7 @@ void loop() {
   
   if (!digitalRead(PIN_BUTTON)) { //Check the BOOTSEL button state (pulled-up --> FALSE=pressed)
     Serial.println("Joystick buttons");
-    for (uint8_t i = 1; i <= 32; i++) {
+    for (int i = 1; i <= 32; i++) {
       Joystick.button(i, true);
       delay(250);
       Joystick.button(i, false);
